+++ src/main/java/org/taxi/pricing/prices/PremiumTaxiRate.java
@@ -17,7 +17,7 @@ public class PremiumTaxiRate implements TaxiRate{
         this.priceCalculator = priceCalculator;
     }
 
-    public double calculatePrice(int distance) {
+    public double calculatePrice(double distance) {
         double price = initialCharge;
         if (distance > tariffBRange){
             price += (tariffB * (distance - tariffBRange));
